+++ lib/turkish_cities/helpers/decomposer_helper.rb
@@ -11,6 +11,10 @@ module DecomposerHelper
     "Couldn't find city name with '#{input}'"
   end
 
+  def cities_not_found_error(first, second)
+    "Couldn't find cities combination with '#{first}/#{second}'"
+  end
+
   def convert_chars(string)
     I18n.transliterate(string.downcase(:turkic))
   end